### 2026-08-08 9b42020
Update GitHub package to align with goodpractice, and pkgcheck mandatory requirements
- Added `.Renviron` and `.Renviron.local` to `.gitignore` to prevent sensitive environment files from being tracked.
- Introduced a new `.Rprofile` file to automatically set the GitHub Personal Access Token (PAT) if not already set, enhancing user experience for package development.

These changes aim to improve the management of environment variables and ensure sensitive information is not exposed in the repository.
