### 2026-08-08 ec33d75
Enhance documentation and improve function exports
- Updated the `.cursorrules` file to clarify the usage of roxygen tags, ensuring that every substantive function is tagged with `@export` for proper documentation and usability.
- Added new internal functions to the NAMESPACE, enhancing the package's functionality and ensuring that they are accessible for internal use.
- Improved error messages in various functions for better clarity and user experience.
- Removed the outdated `.squared_mahalanobis_distance` function from the package, streamlining the codebase.

These changes aim to enhance th
[…]
