### 2026-08-08 bae075d
Update author information
- Revised author names for consistency and corrected formatting in the LaTeX document.
- Added detailed author contributions, conflicts of interest, and acknowledgments sections to improve transparency and recognition of contributions.
- Enhanced the overall structure of the manuscript to align with publication standards.

These changes aim to improve the clarity and professionalism of the manuscript, ensuring proper attribution and acknowledgment of contributions.
