+++ article/main.tex
@@ -48,22 +48,46 @@
     and Multidimensional Framework for Cellular Deconvolution in Heterogeneous Biological
     Samples}
 
-    \author[1,2]{\fnm{Bastien} \sur{CHASSAGNOL}}
+    \author*[1,2]{\fnm{Bastien} \sur{Chassagnol}}
     \email{bastien\_chassagnol@laposte.net}
+    \equalcont{B.C.\ and M.P.\ contributed equally to this work.}
 
-    \author[2,3]{\fnm{Gréogry} \sur{Nuel}}
-    \email{iiauthor@gmail.com}
-    \equalcont{These authors contributed equally to this work.}
+    \author[3]{\fnm{Marielle} \sur{P\'{e}r\'{e}}}
+    \email{marielle.pere@inria.fr}
+    \equalcont{B.C.\ and M.P.\ contributed equally to this work.}
 
-    \author[1,2]{\fnm{Etienne} \sur{Becht}}
-    \email{iiiauthor@gmail.com}
-    \equalcont{These authors contributed equally to this work.}
+    \author[1]{\fnm{Mohamad} \sur{Al Charif}}
 
-    \affil*[1]{\orgdiv{Institut De Recherches Internationales Servier (IRIS), FRANCE}, \orgname{Organization}, \orgaddress{\street{Street}, \city{City}, \postcode{100190}, \state{State}, \country{Country}}}
+    \author[2]{\fnm{Gr\'{e}gory} \sur{Nuel}}
+    \email{nuel@math.cnrs.fr}
 
-    \affil[2]{\orgdiv{LPSM (Laboratoire de Probabilités, Statistiques et Modélisation), Sorbonne Université}, \orgname{Organization}, \orgaddress{\street{Street}, \city{City}, \postcode{10587}, \state{State}, \country{Country}}}
+    \author[4]{\fnm{Etienne} \sur{Becht}}
+    \email{etienne.becht@inserm.fr}
 
-    \affil[3]{\orgdiv{Department}, \orgname{Organization}, \orgaddress{\street{Street}, \city{City}, \postcode{610101}, \state{State}, \country{Country}}}
+    \author[1]{\fnm{Ana\"{i}s} \sur{Baudot}}
+    \email{anais.baudot@univ-amu.fr}
+
+    \affil[1]{\orgdiv{Marseille Medical Genetics (MMG), Systems Biomedicine team},
+      \orgname{Aix-Marseille University, Inserm U1251},
+      \orgaddress{\street{Facult\'{e} de M\'{e}decine, 27 Bd Jean Moulin},
+        \city{Marseille}, \postcode{13385}, \country{France}}}
+
+    \affil[2]{\orgdiv{Laboratoire de Probabilit\'{e}s, Statistique et
+      Mod\'{e}lisation (LPSM), UMR 8001},
+      \orgname{Sorbonne Universit\'{e} / CNRS},
+      \orgaddress{\city{Paris}, \country{France}}}
+
+    \affil[3]{\orgdiv{MacBes team},
+      \orgname{Inria Centre at Universit\'{e} C\^{o}te d'Azur;
+        Institut de Pharmacologie Mol\'{e}culaire et Cellulaire (IPMC)},
+      \orgaddress{\city{Sophia-Antipolis}, \country{France}}}
+
+    \affil[4]{\orgdiv{Centre de Recherche sur l'Inflammation (CRI),
+      UMR 1149},
+      \orgname{Inserm / Universit\'{e} Paris Cit\'{e}},
+      \orgaddress{\street{Facult\'{e} de m\'{e}decine -- site Bichat,
+        16 rue Henri Huchard},
+        \city{Paris}, \postcode{75018}, \country{France}}}
 
     %%================================%%
[…]
