### 2026-08-12 41ad975
Update package documentation and bibliography, to prepare for alignment with srr guidelines, ropensci submission, and address requests from HAL reviewers
- Enhanced the `_pkgdown.yml` file by adding a new article link for "DeCovarT perspectives (outlook)" to improve navigation and accessibility of resources.
- Updated the `.Rbuildignore` file to exclude `B-submit-a-presubmission-inquiry.md` and `pkgcheck-report.md`, ensuring a cleaner build process.
- Modified the `DESCRIPTION` file to include additional roclets in the Roxygen configuration, improving documentation generation.
- Added several 
[…]
