+++ R/srr-stats-standards.R
@@ -1,0 +1,367 @@
+#' @srrstatsVerbose TRUE
+#' @noRd
+NULL
+
+#' srr_stats
+#'
+#' rOpenSci Statistical Software Standards - General (G) and Regression (RE)
+#' categories.  Standards already addressed are tagged `@srrstats`.
+#' Standards not applicable are tagged `@srrstatsNA` in the block below.
+#' Remaining standards keep `@srrstatsTODO` for full review.
+#'
+#' G1 - Documentation and references
+#'
+#' @srrstats {G1.0} Primary reference is Chassagnol, Nuel and Becht (2023)
+#'   <doi:10.48550/arXiv.2309.09557>, cited in DESCRIPTION and
+#'   `inst/REFERENCES.bib`.
+#'
+#' @srrstats {G1.1} DeCovarT is documented as a novel R implementation of
+#'   multivariate Gaussian convolution MLE for bulk deconvolution; no prior
+#'   R implementation of this exact approach exists (README, preprint).
+#'
+#' @srrstats {G1.2} A lifecycle badge (Active/stable,
+#'   <https://www.repostatus.org/#active>) is shown in the README; future
+#'   development is described there and in vignettes.
+#'
+#' @srrstats {G1.3} Statistical terminology (ALR reparametrisation, simplex
+#'   constraint, Fisher information, Marquardt-Levenberg) is defined in the
+#'   vignettes and in `article/main.tex`.
+#'
+#' @srrstats {G1.4} All exported functions are documented with roxygen2
+#'   (title, description, `@param`, `@return`, `@examples`).
+#'
+#' @srrstats {G1.4a} All internal helpers carry a `@noRd` or
+#'   `@keywords internal` tag.
+#'
+#' @srrstatsTODO {G1.5} Reproduction scripts for arXiv preprint performance
+#'   claims to be added to `scripts/` before full submission.
+#'
+#' @srrstatsTODO {G1.6} Comprehensive benchmark vignette comparing DeCovarT
+#'   to alternative implementations is planned.
+#'
+#' G2 - Input validation
+#'
+#' @srrstats {G2.0} Input length assertions implemented via `checkmate`
+#'   (e.g. `assert_matrix`, `assert_numeric`) throughout estimation functions.
+#'
+#' @srrstats {G2.0a} Expected dimensions and lengths documented in `@param`
+#'   of all exported functions.
+#'
+#' @srrstats {G2.1} Type assertions performed via `checkmate` on all inputs
+#'   (numeric matrices, positive-definite covariance matrices, etc.).
+#'
+#' @srrstats {G2.1a} Expected types of all inputs documented in `@param`.
+#'
+#' @srrstats {G2.2} Scalar parameters are checked with `assert_scalar` or
+#'   equivalent.
+#'
+#' @srrstats {G2.3} Character parameters validated with `match.arg()`.
+#'
+#' @srrstats {G2.3a} `match.arg()` used for all fixed-choice character
+#'   arguments (e.g. `method` in `run_deconvolution_benchmark()`).
+#'
+#' @srrstats {G2.3b} Character inputs compared case-insensitively via
+#'   `tolower()` in argument validation.
+#'
+#' @srrstatsTODO {G2.4} Coercion between data types to be reviewed.
+#' @srrstatsTODO {G2.4a} explicit conversion to `integer` via `as.integer()`
+#' @srrstatsTODO {G2.4b} explicit conversion to numeric via `as.numeric()`
+#' @srrstatsTODO {G2.4c} explicit conversion to character via `as.character()`
+#' @srrstatsTODO {G2.4d} explicit conversion to factor via `as.factor()`
[…]
