+++ inst/packages.bib
@@ -1,6 +1,6 @@
 % Auto-generated R / Bioconductor software bibliography.
 % Do NOT edit by hand — regenerate with:
-%   Rscript scripts/generate_packages_bib.R
+%   Rscript scripts/curate_bibliography/generate_packages_bib.R
 % Literature (Zotero) lives in decovart_library.bib / REFERENCES.bib.
 % Quarto YAML should list both:
 %   bibliography:
@@ -41,6 +41,14 @@ Inference},
   year = {2026},
   note = {R package version 5.2.1},
   url = {https://www.bnlearn.com/},
+}
+
+@Manual{R-caret,
+  title = {caret: Classification and Regression Training},
+  author = {Max Kuhn},
+  year = {2024},
+  note = {R package version 7.0-1},
+  url = {https://github.com/topepo/caret/},
 }
 
 @Manual{R-clusterGeneration,
@@ -188,6 +196,14 @@ Marquardt-Levenberg Algorithm},
   url = {https://satijalab.org/seurat},
 }
 
+@Manual{R-tidyr,
+  title = {tidyr: Tidy Messy Data},
+  author = {Hadley Wickham and Davis Vaughan and Maximilian Girlich},
+  year = {2025},
+  note = {R package version 1.3.2},
+  url = {https://tidyr.tidyverse.org},
+}
+
 @Manual{R-tinytable,
   title = {tinytable: Simple and Configurable Tables in HTML, LaTeX, Markdown,
 Word, PNG, PDF, and Typst Formats},
@@ -248,6 +264,18 @@ Word, PNG, PDF, and Typst Formats},
   number = {3},
   pages = {1--22},
   doi = {10.18637/jss.v035.i03},
+}
+
+@Article{caret2008,
+  title = {Building Predictive Models in R Using the caret Package},
+  volume = {28},
+  url = {https://www.jstatsoft.org/index.php/jss/article/view/v028i05},
+  doi = {10.18637/jss.v028.i05},
+  number = {5},
+  journal = {Journal of Statistical Software},
+  author = {{Kuhn} and {Max}},
+  year = {2008},
+  pages = {1–26},
 }
 
 @Article{ComplexHeatmap2016,
@@ -448,4 +476,13 @@ cross-validation fold construction},
   year = {2026},
   note = {R package version 0.0.1},
   url = {https://github.com/VeraPancaldiLab/pipeML},
+}
+
+@Manual{R-skpr,
+  title = {skpr: Design of Experiments Suite: Generate and Evaluate
+Optimal Designs},
+  author = {Tyler Morgan-Wall and George Khoury},
+  year = {2025},
+  note = {R package version 1.9.2},
+  url = {https://CRAN.R-project.org/package=skpr},
 }

[…]
