### 2026-08-12 485f3ca
Fix CI: track man pages for build_* helpers ignored by build* gitignore.
Replace overly broad `build*` with `/build/` so roxygen-generated
man/build_*.Rd files are version-controlled and R CMD check no longer
reports undocumented exports on Windows.

Co-authored-by: Cursor <cursoragent@cursor.com>
