+++ man/build_normalised_precision.Rd
@@ -1,0 +1,29 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/02_02_generate_synthetic_networks.R
+\name{build_normalised_precision}
+\alias{build_normalised_precision}
+\title{Complete a weighted adjacency to an SPD precision}
+\usage{
+build_normalised_precision(weighted_adjacency, precision_shift)
+}
+\arguments{
+\item{weighted_adjacency}{Symmetric numeric matrix (zero diagonal).}
+
+\item{precision_shift}{Positive diagonal cushion \eqn{u}.}
+}
+\value{
+Symmetric positive-definite matrix.
+}
+\description{
+Applies a uniform spectral shift that preserves off-diagonal signs and
+support:
+\deqn{
+  \boldsymbol{\Omega}
+  =
+  \boldsymbol{W}
+  +
+  \bigl(\lvert\lambda_{\min}(\boldsymbol{W})\rvert + u\bigr)
+  \mathbf{I}.
+}
+}
+\keyword{internal}
