+++ man/build_covariance_array_from_precision.Rd
@@ -1,0 +1,23 @@
+% Generated by roxygen2: do not edit by hand
+% Please edit documentation in R/02_02_generate_synthetic_networks.R
+\name{build_covariance_array_from_precision}
+\alias{build_covariance_array_from_precision}
+\title{Build cell-type-specific covariances from precision slices}
+\usage{
+build_covariance_array_from_precision(precision_array)
+}
+\arguments{
+\item{precision_array}{Symmetric positive-definite array
+\eqn{(\boldsymbol{\Omega}_j)_{j}\in\mathcal{M}_{G\times G\times J}}.}
+}
+\value{
+Numeric array of dimension \eqn{G\times G\times J}.
+}
+\description{
+For each cell type \eqn{j}, inverts
+\eqn{\boldsymbol{\Omega}_j} to
+\eqn{\boldsymbol{\Sigma}_j=\boldsymbol{\Omega}_j^{-1}} and stacks the
+result as a \eqn{G\times G\times J} array. Precisions need not be
+shared across cell types.
+}
+\keyword{internal}
