+++ articles/softmax-alr-derivatives.md
@@ -72,10 +72,9 @@ Table 2: Standard algebraic operations induced by a matrix
 
 Package helpers
 [`.bilinear_form()`](https://bastienchassagnol.github.io/DeCovarT/reference/dot-bilinear_form.md)
-and
-[`.squared_mahalanobis_distance()`](https://bastienchassagnol.github.io/DeCovarT/reference/dot-squared_mahalanobis_distance.md)
-implement the middle rows of [Table 2](#tbl-bilinear-forms). The second
-solves \boldsymbol{\Sigma}\boldsymbol{z}=\boldsymbol{\delta} rather than
+and `.squared_mahalanobis_distance()` implement the middle rows of
+[Table 2](#tbl-bilinear-forms). The second solves
+\boldsymbol{\Sigma}\boldsymbol{z}=\boldsymbol{\delta} rather than
 forming \boldsymbol{\Sigma}^{-1} explicitly. Calling
 \boldsymbol{x}^{\mathsf{T}}\boldsymbol{A}\boldsymbol{y} a “dot product”
 is misleading unless \boldsymbol{A}=\boldsymbol{I}; see also the [dot
