+++ B-submit-a-presubmission-inquiry.md
@@ -1,0 +1,399 @@
+# NA
+
+Submitting Author Name: Bastien Chassagnol Submitting Author Github
+Handle: @bastienchassagnol Other Package Authors Github handles: (comma
+separated, delete if none) Repository:
+<https://github.com/bastienchassagnol/DeCovarT> Submission type:
+Pre-submission Language: en
+
+> **ORCID:**
+> [0000-0002-8955-2391](https://orcid.org/0000-0002-8955-2391)
+> **Documentation:** <https://bastienchassagnol.github.io/DeCovarT/>
+> **Method preprint:** <https://arxiv.org/abs/2309.09557>
+
+------------------------------------------------------------------------
+
+- Paste the full DESCRIPTION file inside a code block below:
+
+&nbsp;
+
+    Package: DeCovarT
+    Title: Covariance-Aware Deconvolution of Bulk Transcriptomic Samples
+    Version: 2.0.1
+    Authors@R:
+        person("Bastien", "Chassagnol",
+               email = "bastien_chassagnol@laposte.net",
+               role = c("aut", "cre", "cph"),
+               comment = c(
+                 ORCID = "0000-0002-8955-2391",
+                 LinkedIn = "https://www.linkedin.com/in/bastien-chassagnol-677b67140/?locale=fr"
+               ))
+    Description: Estimates cell-type proportions in bulk transcriptomic samples
+        with a probabilistic convolution model that integrates the gene-gene
+        covariance structure of purified reference profiles. Cellular ratios
+        are recovered by maximum likelihood under a multivariate Gaussian
+        convolution, using analytic gradients and Hessians, an additive
+        log-ratio reparametrisation that enforces the simplex constraint, and
+        Marquardt-Levenberg or Newton-type optimisers. The methodology is
+        described in Chassagnol, Nuel and Becht (2023)
+        <doi:10.48550/arXiv.2309.09557>.
+    License: MIT + file LICENSE
+    Additional_repositories: https://bioconductor.org/packages/release/bioc
+    Encoding: UTF-8
+    Roxygen: list(markdown = TRUE)
+    Suggests:
+        compositions,
+        flextable,
+        ggplot2,
+        knitr,
+        litedown,
+        numDeriv,
+        pkgdown,
+        quarto,
+        reactable (>= 0.4.4),
+        readr,
+        rmarkdown,
+        stringr,
+        testthat (>= 3.0.0),
+        tinytable,
+        withr
+    VignetteBuilder: knitr, quarto
+    Config/testthat/edition: 3
+    Imports:
+        circlize,
+        ComplexHeatmap,
+        dplyr,
+        e1071,
+        igraph,
+        limSolve,
+        marqLevAlg,
+        MASS,
+        methods,
+        Metrics,
+        MixSim,
+        nnls,
+        glmnet,
+        purrr,
+        Rdpack (>= 0.7),
+        rlang,
+        tensor,
+        tibble,
+        tidyr,
+        viridis
+    Depends:
+        R (>= 4.1.0)
+    RdMacros: Rdpack
+    Language: en-GB
+    URL: https://github.com/bastienchassagnol/DeCovarT, https://bastienchassagnol.github.io/DeCovarT/
+    BugReports: https://github.com/bastienchassagnol/DeCovarT/issues
+    Config/roxygen2/version: 8.0.0
+    RoxygenNote: 7.3.3
+
+> *DeCovarT* estimates cell-type proportions in bulk RNA-seq samples
+> using a probabilistic convolution model that incorporates the
+> gene–gene covariance structure of cellular profiles. Ratios are
+> recovered by maximum likelihood under a multivariate Gaussian
[…]
