+++ llms.txt
@@ -53,7 +53,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-07
+2026-08-08
 
 ## 
 
@@ -771,3 +771,5 @@ and glmnet gene scores for signature design.
   derivatives](https://bastienchassagnol.github.io/DeCovarT/articles/softmax-alr-derivatives.md):
 - [OLS reminders for linear
   deconvolution](https://bastienchassagnol.github.io/DeCovarT/articles/DeCovarT-reminders-OLS.md):
+- [DeCovarT
+  perspectives](https://bastienchassagnol.github.io/DeCovarT/articles/decovart-perspectives.md):

[…]
