### 2026-08-12 1aa04dd
Update vignette to add a subsection about integrating the hierarchical cell lineage structure to improve the delineation of closely related cell types
- Added new references to the bibliography files, including key articles on cell type proportion estimation and hierarchical deconvolution, enriching the literature context for users.
- Revised the `decovart-perspectives.qmd` vignette to improve clarity and structure, including an outlook on extending DeCovarT beyond bulk transcriptomic deconvolution and detailing new sections on mixture design and hierarchical deconvolution.

These changes aim 
[…]
