+++ vignettes/decovart-perspectives.qmd
@@ -23,35 +23,19 @@ format:
     toc: true
 ---
 
-::: {.callout-note}
+::: callout-note
 ## Scope
 
-Outlook on extending DeCovarT beyond bulk transcriptomic deconvolution.
-Future sections (spatial transcriptomics, multi-modal / multi-omics) will be
-added here. Compositional reparametrisation is implemented in
-[`additive_logistic()`](../reference/additive_logistic.html) and documented
-numerically in the
-[softmax / ALR derivatives](softmax-alr-derivatives.html) vignette.
+Outlook on extending DeCovarT beyond bulk transcriptomic deconvolution. Sections below cover Scheffé-type mixture extensions, lineage-aware hierarchical deconvolution, and compositional validation; future topics (spatial transcriptomics, multi-modal / multi-omics) will be added here. Compositional reparametrisation is implemented in [`additive_logistic()`](../reference/additive_logistic.html) and documented numerically in the [softmax / ALR derivatives](softmax-alr-derivatives.html) vignette.
 :::
 
-Bulk transcriptomic deconvolution is one instance of a **mixture inverse
-problem**: a composite signal (tissue expression, a spectrum, a chromatogram)
-is observed, and we seek the proportions $\boldsymbol{p}$ of underlying
-components. Because $\boldsymbol{p}$ is **compositional** (parts of a whole),
-estimates must lie on the simplex
-$\Delta^{J-1}=\{\boldsymbol{p}\in[0,1]^J:\sum_j p_j=1\}$.
-[@chassagnolDeCovarTMultidimensionalProbalistic2023] model each reference
-profile as multivariate Gaussian and recover $\boldsymbol{p}$ by constrained
-maximum likelihood; classical deconvolution tools such as `CIBERSORT` assume
-gene-wise independence instead [@newmanRobustEnumerationCell2015].
+Bulk transcriptomic deconvolution is one instance of a **mixture inverse problem**: a composite signal (tissue expression, a spectrum, a chromatogram) is observed, and we seek the proportions $\boldsymbol{p}$ of underlying components. Because $\boldsymbol{p}$ is **compositional** (parts of a whole), estimates must lie on the simplex $\Delta^{J-1}=\{\boldsymbol{p}\in[0,1]^J:\sum_j p_j=1\}$. [@chassagnolDeCovarTMultidimensionalProbalistic2023] model each reference profile as multivariate Gaussian and recover $\boldsymbol{p}$ by constrained maximum likelihood; classical deconvolution tools such as `CIBERSORT` assume gene-wise independence instead [@newmanRobustEnumerationCell2015].
 
 # Extending the generative model {#sec-extending-model}
 
 ## Mixture design and first-order structure {#sec-mixture-design}
 
-In designed mixture experiments (chemistry, food science, formulation),
-component proportions are explicit regressors with no free intercept because
-$\sum_j p_j=1$. A Scheffé cubic model for three components reads
+In designed mixture experiments (chemistry, food science, formulation), component proportions are explicit regressors with no free intercept because $\sum_j p_j=1$. A Scheffé cubic model for three components reads
 
 $$
 y = \beta_1 p_1 + \beta_2 p_2 + \beta_3 p_3
[…]
