+++ article/decovart_library.bib
@@ -156,6 +156,17 @@
   doi = {10.1186/s13059-015-0620-6}
 }
 
+@article{angelXCell20Robust2025,
+  title = {{{xCell}} 2.0: Robust Algorithm for Cell Type Proportion Estimation Predicts Response to Immune Checkpoint Blockade},
+  author = {Angel, Almog and Naom, Loai and {Nabet-Levy}, Shir and Aran, Dvir},
+  year = {2025},
+  journal = {Genome Biology},
+  volume = {26},
+  number = {1},
+  pages = {335},
+  doi = {10.1186/s13059-025-03784-3}
+}
+
 @article{anghelISOpureRImplementationComputational2015,
   title = {{{ISOpureR}}: An {{R}} Implementation of a Computational Purification Algorithm of Mixed Tumour Profiles},
   author = {Anghel, Catalina V and Quon, Gerald and Haider, Syed and Nguyen, Francis and Deshwar, Amit G and Morris, Quaid D and Boutros, Paul C},
@@ -792,6 +803,14 @@
   journal = {Bioinformatics (Oxford, England)},
   volume = {35},
   doi = {10.1093/bioinformatics/btz444}
+}
+
+@misc{ederRectangleRobustScalable2026,
+  title = {Rectangle: Robust and Scalable Multiscale Deconvolution Informed by Single-Cell {{RNA}} Sequencing Data},
+  author = {Eder, Bernhard and Rigato, Irene and Dietrich, Alexander and Merotto, Lorenzo and Sturm, Gregor and Treis, Tim and List, Markus and Theis, Fabian J. and Finotello, Francesca},
+  year = {2026},
+  howpublished = {bioRxiv},
+  doi = {10.64898/2026.07.07.736950}
 }
 
 @article{efronAreSetMicroarrays2009,
@@ -2874,6 +2893,17 @@
   journal = {Frontiers in immunology},
   volume = {6},
   doi = {10.3389/fimmu.2015.00380}
+}
+
+@article{voelklHIDEHierarchicalCelltype2025,
+  title = {{{HIDE}}: Hierarchical Cell-Type Deconvolution},
+  author = {V{\"o}lkl, Dennis and {Mensching-Buhr}, Malte and Sterr, Thomas and Bolz, Sarah and Sch{\"a}fer, Andreas and Seifert, Nicole and Tauschke, Jana and Rayford, Austin and Straume, Oddbj{\o}rn and Zacharias, Helena U and Grellscheid, Sushma Nagaraja and Beissbarth, Tim and Altenbuchinger, Michael and G{\"o}rtler, Franziska},
+  year = {2025},
+  journal = {Bioinformatics},
+  volume = {41},
+  number = {Supplement\_1},
+  pages = {i207-i216},
+  doi = {10.1093/bioinformatics/btaf179}
 }
 
 @misc{wangAccurateEstimationCelltype2022,
