+++ llms.txt
@@ -53,7 +53,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-08
+2026-08-12
 
 ## 
 
