+++ articles/decovart-perspectives.md
@@ -3,7 +3,9 @@
 > **Scope**
 >
 > Outlook on extending DeCovarT beyond bulk transcriptomic
-> deconvolution. Future sections (spatial transcriptomics, multi-modal /
+> deconvolution. Sections below cover Scheffé-type mixture extensions,
+> lineage-aware hierarchical deconvolution, and compositional
+> validation; future topics (spatial transcriptomics, multi-modal /
 > multi-omics) will be added here. Compositional reparametrisation is
 > implemented in
 > [`additive_logistic()`](https://bastienchassagnol.github.io/DeCovarT/reference/additive_logistic.md)
@@ -76,6 +78,147 @@ on \boldsymbol{p} and could still be fitted by constrained MLE, at the
 cost of many additional interaction parameters and careful guard against
 overfitting.
 
+### Lineage-aware hierarchical deconvolution
+
+Flat deconvolution treats every cell type as a sibling on one simplex.
+**Hierarchical** methods instead estimate proportions at several
+granularities linked by a cell-type tree: broad compartments first, then
+closely related subtypes whose mean signatures are collinear at the leaf
+level.
+
+We reuse the DeCovarT set-up from the
+[article](https://bastienchassagnol.github.io/DeCovarT/article/main.pdf):
+gene g=1,\ldots,G; cell type or tree node j; sample i=1,\ldots,N
+(independence across i). Slices follow the usual dot notation
+(\boldsymbol{\mu}\_{\cdot j}, \boldsymbol{y}\_{\cdot i}).
+
+- \boldsymbol{y}=(y\_{gi})\in\mathbb{R}\_{+}^{G\times N}: bulk
+  expression; \boldsymbol{y}\_{\cdot i}\in\mathbb{R}\_{+}^{G} is sample
+  i.
+- \boldsymbol{\mu}=(\mu\_{gj})\in\mathbb{R}^{G\times J}: mean signature;
+  \boldsymbol{\mu}\_{\cdot j}\in\mathbb{R}^{G} is the cross-sample mean
+  profile of type j (as in `MuSiC` ([Wang et al.
+  2019](#ref-wangBulkTissueCell2019))).
+- \boldsymbol{p}=(p\_{ji})\in\\\]0,1\[^{J\times N}: unknown relative
+  proportions; \boldsymbol{p}\_{\cdot i}\in\\\]0,1\[^{J} is the
+  composition of sample i.
+
+\boldsymbol{y}\_{\cdot i}=\boldsymbol{\mu}\\\boldsymbol{p}\_{\cdot i},
+\tag{5}
+
+\sum\_{j=1}^{J} p\_{ji}=1, \qquad p\_{ji}\ge 0 \qquad (i=1,\ldots,N).
+\tag{6}
+
+p\_{j,i}=\sum\_{c\in\mathcal{C}(j)} p\_{c,i} \qquad (i=1,\ldots,N).
+\tag{7}
+
+At each tree level the bulk mixture satisfies
+[Eq. 5](#eq-hierarchical-linear) with each composition on the unit
+simplex ([Eq. 6](#eq-simplex-constraint)). For a parent node j with
+child set \mathcal{C}(j), coherent hierarchical estimates require
+[Eq. 7](#eq-parent-child).
+
+#### Hard versus soft lineage constraints
+
+Despite diverse implementations, hierarchical bulk mRNA deconvolution
+methods fall into **two constraint regimes**:
+
+- **Hard constraints** — child lineage ratios are forced to sum exactly
[…]
