+++ articles/deconvolution-seminar-slides.md
@@ -26,7 +26,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-08
+2026-08-12
 
 ## 
 
