+++ pkgdown.yml
@@ -10,7 +10,7 @@ articles:
   network-inference: network-inference.html
   softmax-alr-derivatives: softmax-alr-derivatives.html
   synthetic-scenarios: synthetic-scenarios.html
-last_built: 2026-08-12T15:17Z
+last_built: 2026-08-15T13:56Z
 urls:
   reference: https://bastienchassagnol.github.io/DeCovarT/reference
   article: https://bastienchassagnol.github.io/DeCovarT/articles
