+++ llms.txt
@@ -53,7 +53,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-12
+2026-08-15
 
 ## 
 
