### 2026-08-18 2d3910f
Enforce proper spelling in the whole package
- Added detailed instructions for testing native Quarto Mermaid before implementing client-side fallbacks in `.cursorrules`.
- Updated `DESCRIPTION` to suggest the `spelling` package for improved documentation quality.
- Corrected terminology from "matricial" to "matrix" in relevant documentation files for consistency and clarity.
