+++ R/03_03_DeCovarT_estimate_ratios_frequentist.R
@@ -468,7 +468,7 @@ hessian_additive_logistic <- function(rho) {
 #' Analytic Hessian
 #' \eqn{\mathbf{H}\in\mathcal{M}_{J\times J}} with entries
 #' \eqn{\mathbf{H}_{i,j}=\partial^{2}\ell/(\partial p_i\partial p_j)},
-#' matching the matricial formulae of the article (quadratic forms in
+#' matching the matrix formulae of the article (quadratic forms in
 #' \eqn{\boldsymbol{\Theta}}, \eqn{\boldsymbol{\Sigma}_i},
 #' \eqn{\boldsymbol{\mu}_{\cdot i}} and residual
 #' \eqn{\boldsymbol{r}=\boldsymbol{y}-\boldsymbol{\mu}\boldsymbol{p}}).
