+++ index.md
@@ -53,7 +53,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-17
+2026-08-18
 
 ## 
 
