### 2026-08-18 #798 @bastienchassagnol
> [@bastienchassagnol](https://github.com/bastienchassagnol)
>
> Please see and address the "✖️" in the checks immediately above.

I'm updating my previous pos -> the srr report is truly comprehensive, and will strongly improve the generalisation ability of DeCovarT -> that said, it will take more time before ingesting all the suggested changes
