### 2026-08-18 fd55de9
First run to remove some of the remaining statistical warnings
- Updated functions to use `.match_arg_ci()` for case-insensitive matching of character inputs, ensuring robustness in user input handling.
- Added detailed documentation for parameters with restricted character inputs, including examples and validation methods.
- Implemented error handling for missing values in key functions, improving the package's reliability and user experience.
- Enhanced vignettes to reflect the new features and provide clearer guidance on the use of the package.
