+++ tests/testthat/helper-srr.R
@@ -1,0 +1,2 @@
+# G3.0: never assert exact equality of floating-point values.
+.tol_srr <- 100 * .Machine$double.eps
