+++ R/04_01_run_benchmark.R
@@ -15,7 +15,12 @@
 #'   from `mu` / the third dimension of `sigma` or `Theta`.
 #' @param second_moment Which second-moment field to require:
 #'   `"sigma"`, `"Theta"`, or `"either"` (default: at least one of
-#'   `sigma` / `Theta`).
+#'   `sigma` / `Theta`). Matching is case-insensitive.
+#'
+#' @srrstats {G2.3} Restricted character input (`second_moment`).
+#' @srrstats {G2.3a} Validated via `.match_arg_ci()` (a `match.arg()`
+#'   equivalent).
+#' @srrstats {G2.3b} Matching is case-insensitive (`tolower()`).
 #'
 #' @return `TRUE` invisibly if the structure is valid; otherwise stops
 #'   with an informative error.
@@ -57,7 +62,7 @@ check_true_theta <- function(
   J = NULL,
   second_moment = c("either", "sigma", "Theta")
 ) {
-  second_moment <- match.arg(second_moment)
+  second_moment <- .match_arg_ci(second_moment, c("either", "sigma", "Theta"))
   if (!is.list(true_theta)) {
     stop("`true_theta` must be a list.", call. = FALSE)
   }
@@ -166,7 +171,11 @@ check_true_theta <- function(
       if (nrow(p_mat) == n_celltypes) {
         # J x N: each column is a sample-wise cellular ratio on the simplex;
         # average across samples for mixture-level summaries.
-        if (any(abs(colSums(p_mat) - 1) > 1e-8) || any(p_mat < 0)) {
+        if (
+          any(abs(colSums(p_mat) - 1) > 100 * .Machine$double.eps) ||
+            any(p_mat < 0) ||
+            anyNA(p_mat)
+        ) {
           stop(
             "`true_theta$p` as J x N must have non-negative columns ",
             "summing to 1.",
@@ -176,7 +185,11 @@ check_true_theta <- function(
         p_vec <- rowMeans(p_mat)
       } else if (ncol(p_mat) == n_celltypes) {
         # N x J: each row is a sample-wise ratio; average across samples.
-        if (any(abs(rowSums(p_mat) - 1) > 1e-8) || any(p_mat < 0)) {
+        if (
+          any(abs(rowSums(p_mat) - 1) > 100 * .Machine$double.eps) ||
+            any(p_mat < 0) ||
+            anyNA(p_mat)
+        ) {
           stop(
             "`true_theta$p` as N x J must have non-negative rows ",
             "summing to 1.",
@@ -192,7 +205,11 @@ check_true_theta <- function(
       }
     }
     # Final simplex check on the (possibly averaged) length-J weight vector.
-    if (any(p_vec < 0) || abs(sum(p_vec) - 1) > 1e-8) {
+    if (
+      any(p_vec < 0) ||
+        anyNA(p_vec) ||
+        abs(sum(p_vec) - 1) > 100 * .Machine$double.eps
+    ) {
       stop(
         "`true_theta$p` must be non-negative and sum to 1 ",
         "(after averaging over samples if matrix).",
@@ -360,6 +377,23 @@ compute_benchmark_metrics <- function(
 #'   ),
 #'   cores = 1
 #' )
+#' @srrstats {G1.0} Closest published statistical method is DSection
+#'   (Erkkila et al. 2010): univariate Gaussian convolution. DeMix, DeMixT
+#'   and ISOpureR also treat cell-type profiles as latent; Ogundijo and Wang
+#'   (2017) extend DSection by sequential Monte Carlo. DeCovarT generalises
+#'   DSection to a multivariate convolution with sparse cell-type covariance.
[…]
