### 2026-08-19 c022f7c
Second, and final validation of the srr G tags
- Deleted the `contaminations - semi-referenced based.bib` file as it is no longer needed.
- Removed the `00_01_input_checks.R` file, consolidating input validation functions into `R/utils.R`.
- Introduced new utility functions in `R/utils.R` for case-insensitive argument matching and error handling for missing or non-finite values.
- Updated the `deconvolute_ratios` function to leverage the new input validation utilities, improving robustness and clarity.
- Added optional file output functionality to the `plot_correlation_Heatmap` function, allow
[…]
