+++ R/utils.R
@@ -1,0 +1,408 @@
+#' Case-insensitive match to a set of allowed strings
+#'
+#' Like [match.arg()], but `arg` is matched after `tolower()` so `"Sigma"`
+#' and `"sigma"` are equivalent. If `arg` is the full `choices` vector (the
+#' usual default when the caller omits the argument), the first choice is
+#' returned.
+#'
+#' @param arg Character vector supplied by the user.
+#' @param choices Character vector of allowed values (canonical case).
+#'
+#' @return A character vector of canonical `choices` entries.
+#'
+#' @srrstats {G2.3} Univariate character arguments are restricted to a
+#'   documented set of values.
+#' @srrstats {G2.3a} Validation uses this `match.arg()` equivalent.
+#' @srrstats {G2.3b} Matching is case-insensitive via `tolower()`.
+#' @srrstats {G3.0} Numeric "equality" elsewhere uses
+#'   `100 * .Machine$double.eps`, not `==` on floats.
+#'
+#' @keywords internal
+#' @noRd
+.match_arg_ci <- function(arg, choices) {
+  if (!is.character(arg) || length(arg) < 1L) {
+    stop("Argument must be a non-empty character vector.", call. = FALSE)
+  }
+  if (anyNA(arg)) {
+    stop("Argument must not contain missing values.", call. = FALSE)
+  }
+  # match.arg() default: omitted arg is the full choices vector.
+  if (
+    length(arg) == length(choices) &&
+      setequal(tolower(arg), tolower(choices))
+  ) {
+    arg <- arg[[1L]]
+  }
+  vapply(
+    arg,
+    function(a) {
+      hit <- choices[tolower(choices) == tolower(a)]
+      if (length(hit) != 1L) {
+        stop(
+          "Argument must be one of ",
+          paste(choices, collapse = ", "),
+          ".",
+          call. = FALSE
+        )
+      }
+      hit[[1L]]
+    },
+    character(1L),
+    USE.NAMES = FALSE
+  )
+}
+
+#' Error if an object contains NA, NaN, Inf or -Inf
+#'
+#' Transcriptomic counts after quantification are complete matrices:
+#' zeros (dropouts) are valid observations, but missing or non-finite
+#' values are not. Proteomic intensities, by contrast, often contain
+#' structurally missing entries (limit of detection, DDA sampling);
+#' that case is out of scope for the current API.
+#'
+#' @param x Atomic vector, matrix, or array.
+#' @param name Argument name used in the error message.
+#'
+#' @srrstats {G2.13} Missing values are rejected before analytic routines.
+#' @srrstats {G2.14} The only supported missing-data policy is to error
+#'   (no ignore / impute options).
+#' @srrstats {G2.14a} `NA` / `NaN` raise an error.
+#' @srrstats {G2.15} Downstream `mean` / `var` / `cor` therefore never see
+#'   incomplete data (they keep default `na.rm = FALSE`).
+#' @srrstats {G2.16} `Inf` and `-Inf` are rejected as well.
+#'
+#' @keywords internal
+#' @noRd
+.assert_no_missing <- function(x, name) {
+  if (anyNA(x)) {
+    stop(
+      "`",
+      name,
+      "` must not contain missing or NaN values.",
+      call. = FALSE
+    )
+  }
+  if (any(!is.finite(x))) {
+    stop("`", name, "` must not contain Inf or -Inf.", call. = FALSE)
+  }
+  invisible(TRUE)
+}
+
[…]
