+++ reference/deconvolute_ratios.md
@@ -29,22 +29,27 @@ deconvolute_ratios(
 - signature_matrix:
 
   Mean signature \\\boldsymbol{\mu}\in\mathcal{M}\_{G\times J}\\
-  (rownames = genes, colnames = cell types). Used as a frequentist
-  **plug-in** for unobserved latent cell-type profiles
-  \\\boldsymbol{x}\_{\cdot j}\\.
+  (rownames = genes, colnames = cell types). Numeric matrix of
+  non-negative expression; \\G \ge J\\ is required (the mixture is
+  undetermined if \\J \> G\\). Used as a frequentist **plug-in** for
+  unobserved latent cell-type profiles \\\boldsymbol{x}\_{\cdot j}\\.
 
 - bulk_expression:
 
-  Bulk matrix \\\boldsymbol{Y}\in\mathcal{M}\_{G\times N}\\.
+  Bulk matrix \\\boldsymbol{Y}\in\mathcal{M}\_{G\times N}\\: numeric,
+  non-negative, gene rownames matching `signature_matrix`.
 
 - true_ratios:
 
-  Optional ground-truth proportions for scoring.
+  Optional ground-truth proportions: a length-\\J\\ numeric vector
+  (recycled over samples) or a numeric matrix \\J\times N\\ (or
+  \\N\times J\\).
 
 - Sigma:
 
   Optional array \\(\boldsymbol{\Sigma}\_j)\_{j}\in\mathcal{M}\_{G\times
-  G\times J}\\.
+  G\times J}\\ of cell-type covariances (numeric; off-diagonals may be
+  negative).
 
 - deconvolution_functions:
 
@@ -63,6 +68,38 @@ deconvolute_ratios(
 
 A `tibble` of estimated \\\hat{\boldsymbol{p}}\\ and metrics, after
 [`repair_simplex()`](https://bastienchassagnol.github.io/DeCovarT/reference/repair_simplex.md).
+
+## References
+
+Erkkilä T, Lehmusvaara S, Ruusuvuori P, Visakorpi T, Shmulevich I,
+Lähdesmäki H (2010). “Probabilistic Analysis of Gene Expression
+Measurements from Heterogeneous Tissues.” *Bioinformatics*, **26**.
+[doi:10.1093/bioinformatics/btq406](https://doi.org/10.1093/bioinformatics/btq406)
+. Ahn J, Yuan Y, Parmigiani G, Suraokar MB, Diao L, Wistuba II, Wang W
+(2013). “DeMix: Deconvolution for Mixed Cancer Transcriptomes Using Raw
+Measured Data.” *Bioinformatics (Oxford, England)*, **29**.
+[doi:10.1093/bioinformatics/btt301](https://doi.org/10.1093/bioinformatics/btt301)
+. Wang Z, Cao S, Morris JS, Ahn J, Liu R, Tyekucheva S, Gao F, Li B, Lu
+W, Tang X, Wistuba II, Bowden M, Mucci L, Loda M, Parmigiani G, Holmes
+CC, Wang W (2018). “Transcriptome Deconvolution of Heterogeneous Tumor
+Samples with Immune Infiltration.” *iScience*, **9**.
+[doi:10.1016/j.isci.2018.10.028](https://doi.org/10.1016/j.isci.2018.10.028)
+. Anghel CV, Quon G, Haider S, Nguyen F, Deshwar AG, Morris QD, Boutros
+PC (2015). “ISOpureR: An R Implementation of a Computational
+Purification Algorithm of Mixed Tumour Profiles.” *BMC Bioinformatics*,
+**16**.
+[doi:10.1186/s12859-015-0597-x](https://doi.org/10.1186/s12859-015-0597-x)
[…]
