### 2026-08-19 9714c7f
Enhance package functionality and documentation, by udpating the perspevtive litterature, and adding a wrapper to deconvolute_ratios
- Added `fit_decovart()` S3 API, providing a dedicated convolution fit object with methods for `print`, `summary`, `coef`, `fitted`, `residuals`, `vcov`, `nobs`, and `confint`.
- Introduced gene-wise affine standardisation for bulk, means, and covariances, improving the handling of data.
- Updated the `deconvolute_ratios()` function to include a `return_model` parameter, allowing users to retrieve model diagnostics alongside estimated proportions.
- Enhanced the 
[…]
