+++ tests/testthat/test-03_05_decovart_fit.R
@@ -1,0 +1,249 @@
+test_that("fit_decovart S3 accessors match the convolution (RE4, RE6)", {
+  toy <- .toy_deconvolution()
+  fit <- suppressWarnings(fit_decovart(
+    signature_matrix = toy$signature_matrix,
+    bulk_expression = toy$bulk_expression,
+    true_ratios = toy$true_ratios,
+    Sigma = toy$Sigma,
+    method = "Marquardt-Levenberg",
+    itmax = 60
+  ))
+
+  expect_s3_class(fit, "decovart_fit")
+  expect_identical(rownames(coef(fit)), colnames(toy$signature_matrix))
+  expect_identical(colnames(coef(fit)), colnames(toy$bulk_expression))
+  expect_equal(
+    fitted(fit),
+    toy$signature_matrix %*% coef(fit),
+    ignore_attr = TRUE,
+    tolerance = .tol_srr
+  )
+  expect_equal(
+    residuals(fit),
+    toy$bulk_expression - fitted(fit),
+    ignore_attr = TRUE,
+    tolerance = .tol_srr
+  )
+  n <- nobs(fit)
+  expect_identical(as.integer(n), ncol(toy$bulk_expression))
+  expect_identical(attr(n, "n_genes"), nrow(toy$signature_matrix))
+  expect_identical(attr(n, "n_celltypes"), ncol(toy$signature_matrix))
+  expect_true(is.matrix(vcov(fit)) || is.list(vcov(fit)))
+  expect_output(print(fit), "DeCovarT fit")
+  expect_output(print(summary(fit)), "log-likelihood")
+  grDevices::pdf(NULL)
+  expect_silent(plot(fit))
+  grDevices::dev.off()
+  expect_false(is.function(utils::getS3method(
+    "predict",
+    "decovart_fit",
+    optional = TRUE
+  )))
+  expect_false(is.function(utils::getS3method(
+    "formula",
+    "decovart_fit",
+    optional = TRUE
+  )))
+})
+
+test_that("gene-wise affine maps leave the MLE unchanged (RE2.3)", {
+  genes <- paste0("g", 1:3)
+  cts <- paste0("ct", 1:2)
+  mu <- matrix(
+    c(20, 40, 25, 35, 30, 22),
+    nrow = 3,
+    dimnames = list(genes, cts)
+  )
+  Sigma <- array(0, dim = c(3, 3, 2), dimnames = list(genes, genes, cts))
+  Sigma[,, 1] <- diag(c(1, 1.2, 0.8))
+  Sigma[,, 2] <- diag(c(1.5, 0.9, 1.1))
+  y <- matrix(
+    drop(mu %*% c(0.65, 0.35)),
+    ncol = 1,
+    dimnames = list(genes, "s1")
+  )
+
+  p_raw <- suppressWarnings(deconvolute_ratios_Marquardt_Levenberg(
+    drop(y),
+    mu,
+    Sigma,
+    itmax = 80
+  ))
+  centre <- c(2, -1, 0.5)
+  scale <- c(2, 0.5, 4)
+  y_star <- (y - centre) / scale
+  mu_star <- (mu - centre) / scale
+  Sigma_star <- Sigma
+  for (j in seq_len(2)) {
+    Sigma_star[,, j] <- Sigma[,, j] * tcrossprod(1 / scale)
+  }
+  p_star <- suppressWarnings(deconvolute_ratios_Marquardt_Levenberg(
+    drop(y_star),
+    mu_star,
[…]
