### 2026-08-19 2a8c665
Release 2.2.0: document helpers, tidy goodpractice lintr, embed the methods paper.
Address the rOpenSci lintr items (unnested pipes, expect_length/expect_type, toString), add Rd examples and inheritParams, and publish article/main.pdf on the pkgdown site via embedpdf.

Co-authored-by: Cursor <cursoragent@cursor.com>
