### 2026-08-20 3f7e936
Prepare 2.2.0 for CRAN: first-release hygiene and ThinkR extra checks.
Shorten the attach message, drop the LinkedIn URL, ship MIT install docs, and keep the perspectives article off the tarball so incoming checks stay clean.

Co-authored-by: Cursor <cursoragent@cursor.com>
