+++ reference/loglik_multivariate_constrained.md
@@ -37,3 +37,14 @@ loglik_multivariate_constrained(rho, y, mean_signature_matrix, Sigma)
 ## Value
 
 Scalar log-likelihood on the constrained manifold.
+
+## Examples
+
+``` r
+mu <- matrix(c(20, 22, 22, 20), 2)
+Sigma <- array(c(diag(2), diag(2)), dim = c(2, 2, 2))
+p <- c(0.6, 0.4)
+y <- drop(mu %*% p)
+loglik_multivariate_constrained(additive_log_ratio(p), y, mu, Sigma)
+#> [1] 1.307853
+```
