+++ reference/generate_random_network_skeleton.md
@@ -55,3 +55,18 @@ generate_random_network_skeleton(n_genes, graph_model, graph_params = list())
 ## Value
 
 Symmetric integer matrix \\G\times G\\ with zero diagonal.
+
+## Examples
+
+``` r
+generate_random_network_skeleton(8L, "erdos_renyi")
+#>      [,1] [,2] [,3] [,4] [,5] [,6] [,7] [,8]
+#> [1,]    0    0    0    0    0    0    0    1
+#> [2,]    0    0    0    0    0    0    1    0
+#> [3,]    0    0    0    0    1    0    0    0
+#> [4,]    0    0    0    0    1    1    0    0
+#> [5,]    0    0    1    1    0    1    0    0
+#> [6,]    0    0    0    1    1    0    0    1
+#> [7,]    0    1    0    0    0    0    0    1
+#> [8,]    1    0    0    0    0    1    1    0
+```
