+++ reference/dot-sigma_p_factorisation.md
@@ -54,3 +54,12 @@ only the first of those calls actually factorises, and the rest simply
 return the cached result in \\O(G^{2})\\ (the cost of the equality
 check). Profiling on a 38-gene / 3-cell-type scenario showed the
 redundancy
+
+## Examples
+
+``` r
+p <- c(0.6, 0.4)
+Sigma <- array(c(diag(2), diag(2)), dim = c(2, 2, 2))
+names(.sigma_p_factorisation(p, Sigma))
+#> [1] "matrix"  "chol"    "log_det" "inverse"
+```
