+++ reference/hessian_additive_logistic.md
@@ -20,3 +20,27 @@ hessian_additive_logistic(rho)
 ## Value
 
 Numeric array used in the constrained Hessian chain rule.
+
+## Examples
+
+``` r
+hessian_additive_logistic(c(0.2, -0.5))
+#> , , 1
+#> 
+#>             [,1]        [,2]
+#> [1,]  0.03341530 -0.01261563
+#> [2,] -0.01261563 -0.05289836
+#> 
+#> , , 2
+#> 
+#>             [,1]        [,2]
+#> [1,] -0.01261563 -0.05289836
+#> [2,] -0.05289836  0.09620788
+#> 
+#> , , 3
+#> 
+#>             [,1]        [,2]
+#> [1,] -0.02079966  0.06551400
+#> [2,]  0.06551400 -0.04330952
+#> 
+```
