+++ reference/loglik_multivariate.md
@@ -61,3 +61,20 @@ step (see
 
 [`gradient_loglik_unconstrained()`](https://bastienchassagnol.github.io/DeCovarT/reference/gradient_loglik_unconstrained.md),
 [`additive_logistic()`](https://bastienchassagnol.github.io/DeCovarT/reference/additive_logistic.md)
+
+## Examples
+
+``` r
+genes <- paste0("g", 1:2)
+cts <- paste0("ct", 1:2)
+mu <- matrix(c(20, 22, 22, 20), nrow = 2, dimnames = list(genes, cts))
+Sigma <- array(
+  c(1, 0, 0, 1, 1, 0, 0, 1),
+  dim = c(2, 2, 2),
+  dimnames = list(genes, genes, cts)
+)
+p <- c(0.6, 0.4)
+y <- drop(mu %*% p)
+loglik_multivariate(p, y, mu, Sigma)
+#> [1] 1.307853
+```
