+++ reference/dot-map_gaussian_convolution.md
@@ -19,18 +19,44 @@ estimating proportions alone.
 
 - y:
 
-  Bulk vector \\\boldsymbol{y}\in\mathbb{R}^{G}\\.
+  Numeric vector (or one-column matrix)
+  \\\boldsymbol{y}\in\mathbb{R}^{G}\\.
 
 - mean_signature_matrix:
 
-  Mean matrix \\\boldsymbol{\mu}\in\mathcal{M}\_{G\times J}\\ (prior /
-  plug-in means).
+  Numeric matrix \\\boldsymbol{\mu}\in\mathcal{M}\_{G\times J}\\
+  (plug-in means).
 
 - Sigma:
 
-  Array \\(\boldsymbol{\Sigma}\_j)\_{j}\in\mathcal{M}\_{G\times G\times
+  Array of cell-type covariances in \\\mathcal{M}\_{G\times G\times
   J}\\.
 
 ## Value
 
 List of length \\J\\ with MAP vectors in \\\mathbb{R}^{G}\\.
+
+## Examples
+
+``` r
+genes <- paste0("g", 1:2)
+cts <- paste0("ct", 1:2)
+mu <- matrix(c(20, 22, 22, 20), nrow = 2, dimnames = list(genes, cts))
+Sigma <- array(
+  c(1, 0, 0, 1, 1, 0, 0, 1),
+  dim = c(2, 2, 2),
+  dimnames = list(genes, genes, cts)
+)
+y <- drop(mu %*% c(0.6, 0.4))
+.map_gaussian_convolution(y, mu, Sigma)
+#> [[1]]
+#>     ct1
+#> g1  9.4
+#> g2 11.6
+#> 
+#> [[2]]
+#>     ct2
+#> g1 11.4
+#> g2  9.6
+#> 
+```

[…]
