### 2026-08-20 fe1fa1f
Bump version to 2.2.1 [ci skip]
