### 2026-08-21 137ac26
Update package for CRAN submission: bump version to 2.2.3, remove `Additional_repositories`, and move heatmap dependencies to Suggests. Enhance documentation for optional heatmap functionality and ensure Quarto vignettes use Markdown tables to avoid dependency issues during rebuilds.
