+++ vignettes/DeCoVart-use-cases.qmd
@@ -23,12 +23,6 @@ format:
     toc: true
 ---
 
-```{r}
-#| label: setup-use-cases
-#| include: false
-options(tinytable_html_mathjax = TRUE)
-```
-
 This vignette walks through two complementary simulation use cases for `DeCovarT`:
 
 1. a **bivariate toy model** ($G=2$ genes, $J=2$ cell types) that isolates the effect of gene–gene correlation on mean-only versus covariance-aware deconvolution;
@@ -36,7 +30,7 @@ This vignette walks through two complementary simulation use cases for `DeCovarT
 
 # Bivariate toy model: when covariance hurts mean-only deconvolution {#sec-bivariate-toy}
 
-The manuscript's bivariate study (two genes, two cell populations) is implemented by [`benchmark_bivariate_gaussian_convolutions()`](../reference/benchmark_bivariate_gaussian_convolutions.html) in `R/02_01_toy_simulation.R`. The purpose is deliberately narrow: show that **even when mean expression profiles are similar**, accounting for the covariance between genes can (a) **degrade** standard mean-only deconvolution algorithms, and (b) be **partly alleviated** by DeCovarT's native inclusion of $\boldsymbol{\Sigma}(\boldsymbol{p})=\sum_j p_j^2\boldsymbol{\Sigma}_j$.
+The manuscript's bivariate study (two genes, two cell populations) is implemented by `benchmark_bivariate_gaussian_convolutions()` in `R/02_01_toy_simulation.R`. The purpose is deliberately narrow: show that **even when mean expression profiles are similar**, accounting for the covariance between genes can (a) **degrade** standard mean-only deconvolution algorithms, and (b) be **partly alleviated** by DeCovarT's native inclusion of $\boldsymbol{\Sigma}(\boldsymbol{p})=\sum_j p_j^2\boldsymbol{\Sigma}_j$.
 
 ## Generative design
 
@@ -87,7 +81,7 @@ The returned list has two tibbles: `config` (design, Shannon entropy of $\boldsy
 
 # High-dimensional hybrid scenario and deconvolution {#sec-scenario-grid}
 
-The earlier factorial design swept one topology and one cosine level at a time, keeping a single **shared** covariance across all $J$ cell types (`simulate_hierarchical_grn_moments()`; see [GGM networks](synthetic-scenarios.html#sec-ggm-networks)). `scripts/generate_random_markov_network.R` instead builds **one** scenario that stresses several axes of the feature-selection pipeline simultaneously: two cell types that are unrecoverable from $\boldsymbol{\mu}$ alone and must be told apart from **topology**, a third cell type set apart by a compact marker block, and a null block that should survive every selection stage as *discarded*. Notation matches the manuscript: $G=50$ genes, $J=3$ cell types, $N$ bulk samples for $\boldsymbol{Y}\in\mathcal{M}_{G\times N}$.
[…]
