+++ man/compute_shannon_entropy.Rd
@@ -45,7 +45,7 @@ Panel B of the figure contrasts a peaked (type-specific, low entropy)
 share vector with a flat (multi-type, high entropy) one for \eqn{J=5}
 classes.
 
-\if{html}{\figure{gini_vs_entropy_specificity.png}{options: width="100\%"
+\if{html}{\figure{gini_vs_entropy_specificity.png}{options: width=700
   alt="Gini versus Shannon entropy"}}
 \if{latex}{\figure{gini_vs_entropy_specificity.png}{options: width=5.5in}}
 }

[…]
