+++ llms.txt
@@ -53,7 +53,7 @@ Etienne Becht
 
 Anaïs Baudot
 
-2026-08-20
+2026-08-21
 
 ## 
 
