### 2026-08-23 6db4723
Add Graphify knowledge graph integration and update documentation
- Introduced Graphify knowledge graph support, including rules for usage and commands for exploring codebase dependencies.
- Added `.graphifyignore` to manage extraction scope and prevent unnecessary files from being included in the graph.
- Updated `.gitignore` to exclude the `graphify-out/` directory, ensuring it remains local and is rebuilt as needed.
- Created `.gitattributes` to define merge strategies for the Graphify output files.

This enhancement improves code exploration and understanding of dependencies within the DeC
[…]
