+++ derive-statisticazl-properties.md
@@ -1,0 +1,1652 @@
+# DeCovarT likelihood geometry, constrained inference, and rOpenSci regression standards
+
+## Principal conclusion
+
+The central theoretical claim you proposed — that the DeCovarT sample log-likelihood is globally concave and therefore has a unique global maximum because its Hessian is everywhere strictly negative definite — is **false, even under the favourable assumptions you specified**: strictly positive expression values, strictly positive-definite cell-type covariance matrices, \(G>J\), linearly independent signature columns, and true proportions strictly inside the simplex.
+
+There are, however, two strong results that *can* be proved and are, in my view, the right theoretical statements for the article:
+
+1. **Global identifiability and a unique population optimum.** If the signature matrix is injective on the simplex — full column rank of \(\mu\) is a sufficient, though stronger-than-necessary condition — then
+   \[
+   p\longmapsto
+   \left(\mu p,\ \Sigma(p)\right),
+   \qquad
+   \Sigma(p)=\sum_jp_j^2\Sigma_j ,
+   \]
+   is identifiable. Consequently the expected log-likelihood under the true \(p_0\) has a **unique global maximum at \(p_0\)**, because its difference from the optimum is a Kullback–Leibler divergence.
+
+2. **Strict local curvature in expectation.** For the correctly normalised Gaussian likelihood, the expected Fisher information is strictly positive definite under your full-rank assumption. Hence the **expected Hessian is strictly negative definite** in identifiable directions:
+   \[
+   \mathbb E_p[-H_p(Y)]=I_p\succ0.
+   \]
+   After restriction to the simplex or transformation to ALR coordinates, the corresponding Fisher information remains strictly positive definite.
+
+These are considerably stronger and cleaner results than merely saying that an optimiser usually converges, but they must not be confused with global concavity of the realised, sample-specific likelihood.
+
+More strongly, there is an explicit \(J=2,G=3\) counterexample satisfying all your assumptions in which the correctly specified DeCovarT likelihood has **two distinct, equal, interior global maxima**. The ALR transformation preserves those two maxima. Thus neither positivity, covariance positive-definiteness, nor absence of collinearity can establish uniqueness of the finite-sample MLE.
+
+Before developing that theory, there is also an important implementation/manuscript correction: the Gaussian determinant term in both the current article and the current R implementation is off by a factor of two. Your model is stated as
+\[
+Y\mid p\sim N_G\!\left(\mu p,\Sigma(p)\right),
+\]
+but the current code evaluates
+\[
+-\log\det\Sigma(p)
+-\frac12r^\top\Sigma(p)^{-1}r,
+\]
+whereas the multivariate-normal log-density requires
+\[
+-\frac12\log\det\Sigma(p)
+-\frac12r^\top\Sigma(p)^{-1}r.
+\]
[…]
