### 2026-08-24 #798 @mpadge
@bastienchassagnol Thanks for your very detailed and helpful responses. Your comments on algorithmic benchmarking and scaling behaviour make sense, and seem to provide a good argument for **not** complying with those standards here. The compliance report above shows very good compliance without those, so we'd be happy to proceed to a full review.

Before we do so, I do note that a few  of your comments above suggest you are still updating compliance, or changing code to comply. We'd first ask you to ensure any changes are fully implemented first, so the package is in a stable state when you open a full review issue.

And one further concern that I suspect may arise during review is your statement that your log-likelihood is not globally concave. That seems like the kind of behaviour that will definitely need to be thoroughly explored, documented, and tested before proceeding. Non-concavity is also naturally very common, and I imagine something like your suggested `warn_convergence` flag will be a great way to handle it. But please ensure that all such changes have been implemented, documented, and tested before proceeding.

----

With that, you may close this issue any time you like, and we invite you to open a full submission issue once the package has reached a stable state. Thank you 👍
