### 2026-08-24 #798 @mpadge
One further comment @bastienchassagnol: Thank you very much for noting the "_Not yet in scope_" disclaimer on probability distributions. That was an oversight on our part which has [now been fixed](https://github.com/ropensci/statistical-software-review-book/commit/974cd8f0d73961235c74bfa34b78086d39fd8817). Probability distributions are definitely in scope, and your compliance with them is very welcome!
