### 2026-08-25 #798 @bastienchassagnol
@mpadge , and @maurolepore

 Thank you for all your comments, they were very helpful, and I am glad that DeCovarT is close to being acceptable for full review.

On algorithmic benchmarking: thanks for agreeing that a  comprehensive benchmark against existing mean-only deconvolution tools is not required here. DeCovarT embodies a genuine paradigm shift by modelling dependencies between transcript abundances (covariance / precision structure), so a comprehensive head-to-head comparison would:
   (i) impose a large computational burden and
   (ii) remain hard to interpret as a fair comparison, because competing methods typically need only cell-type means whereas DeCovarT additionally depends on reconstructed second-moment structure (sparse precision matrices preferentially). I will keep that dependence explicit in the documentation rather than claiming a universal accuracy ranking against mean-only pipelines.

On the other hand, before opening a full review issue, I will account for your second set of comments, bringing the package to a **stable** state, by:
  1. **Adding extra tests** for stability under small perturbations of the initial estimate (multi-start / nearby ALR starts), and for identifiable vs near-collinear / multimodal regimes.
  2. **Document and test non-global concavity** of the finite-sample log-likelihood: the population criterion can be uniquely maximised under standard assumptions, but the realised sample objective need not be globally concave and can be multimodal even in simple \(J=2\) settings. I will add a theoretical vignette on the impact of that geometry on parameter recovery and interval coverage, and expose clearer convergence / multi-start diagnostics.
  3. I wil additionally document **Probability Distributions** standards (PD) in the `srr` report, now that the “not yet in scope” disclaimer has been corrected — thank you for confirming that compliance there is welcome.

With that, I am **closing this pre-submission inquiry**. Your comments show a clear interest in both the package and its statistical implementation, which I greatly appreciate. I aim to open a **full submission / review issue by the end of September at the latest**, once the stability, non-concavity documentation/tests are in place.

Thanks again for the insightful guidance.
