### 2026-08-25 #798 @mpadge
Thank you very much @bastienchassagnol, We'll look forward to receiving your full submission. And since you're diving very deeply into covariance structures here, please also feel very welcome to open issues or pull requests on any aspects of our [statistical standards](https://github.com/ropensci/statistical-software-review-book) which you think may be able to be improved. We strongly encourage community feedback there.
